Make better flash sequences.
These flashes are not perfect - they jitter a fair bit at low
brightness - but they are better than previously.

This is probably the best we can do with 50Hz PWM and 20 pulse widths
(ie 100Hz interrupts.)  If the ISR can be made shorter, the PWM
frequency could be increased, or the pulse widths decreased to fix
this problem.

Signed-off-by: Russell Steicke <russells@adelie.cx>