Count the auto-off timeout correctly.
Move the timeout counter into the timer interrupt, so we count the
timeout correctly.

Previously, with the timeout at 15 minutes, it took about 17 minutes
to turn off.  We were missing a significant number of interrupts
(about 13%), which badly affected the timeout.  Now we get a fairly
accurate timeout, affected mainly by the accuracy of the inbuilt RC
oscillator.

Also, correct the timeout for the new higher interrupt rate.

Signed-off-by: Russell Steicke <russells@adelie.cx>