Fix the button checks at startup.
The button press check was completely inverted, creating a bug where
if the button was pressed for too long at startup, we would
immediately switch off when the button was released.  The code thought
that the button state was opposite to what it really was.  Because of
the requirement that the button be released for three cycles before we
start to check for another press, it sort of worked as it was.

We now correctly check for button press and release, and the behaviour
on button presses is reliable.

Signed-off-by: Russell Steicke <russells@adelie.cx>